% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_day_items.R
\docType{data}
\name{everyday_items}
\alias{everyday_items}
\title{Dimensions of everyday items.}
\format{
A data frame with 5 variables:
\describe{
 \item{what}{the name}
 \item{height}{meters in the y direction usually}
 \item{width}{meters in the x direction usually}
 \item{depth}{meters in the z direction usually}
 \item{type}{what kind of thing}
}
}
\source{
\url{https://www.dimensions.guide/}
}
\usage{
everyday_items
}
\description{
Containing a few animals, types of cars and devices.
All from dimensions.guide
With units size. The variables all are in meters.
}
\keyword{datasets}
